… aten/src/ATen/cuda/LegacyTHFunctionsCUDA.cpp
… aten/src/ATen/native/LegacyNNDefinitions.cpp
… aten/src/THCUNN/SpatialDepthwiseConvolution.cu
… aten/src/THCUNN/generic/SpatialDepthwiseConvolution.cu