@ejguan 2021-07-27
Reverting this PR since it breaks `check_backward_compatibility`. Please add BC-breaking changes to the allow list  in https://github.com/pytorch/pytorch/blob/master/test/backward_compatibility/check_backward_compatibility.py